ref(scim): Move SCIM endpoints and tests to core folder
- Moved SCIM endpoint files from src/sentry/scim/endpoints/ to src/sentry/core/endpoints/scim/
- Moved corresponding test files from tests/sentry/api/endpoints/ to tests/sentry/core/endpoints/scim/
- Updated import references in src/sentry/api/urls.py and related serializer files
- Removed empty src/sentry/scim/endpoints/ directory
