+++ src/sentry/core/endpoints/scim/schemas.py
