+++ tests/sentry/core/endpoints/scim/test_scim_team_details.py
