+++ tests/sentry/core/endpoints/scim/test_scim_user_details.py
